// app/src/store/getter-types.ts
export enum GetterTypes {
  EXAMPLE_WITH_ARG = 'EXAMPLE_WITH_ARG',

  TREE_CONFIG = 'TREE_CONFIG',

  CURRENT_S3_BUCKET = 'CURRENT_S3_BUCKET',
  S3_READY = 'S3_READY',

  PARENT_FOLDER = 'PARENT_FOLDER',

  ELEMENT_INITIAL = 'ELEMENT_INITIAL',
  ELEMENT_LOADING = 'ELEMENT_LOADING',
  ELEMENT_SUCCESS = 'ELEMENT_SUCCESS',
  ELEMENT_ERROR = 'ELEMENT_ERROR',
  ELEMENT_DISABLED = 'ELEMENT_DISABLED',
  ELEMENT_STATUS_MAP = 'ELEMENT_STATUS_MAP',
}